Add `FromUTF8View`, `Substring`/`UTF8View` inference overloads
While we have `Parsers.Pullback`, which makes it easy to
`.pullback(\.utf8)` any `UTF8View`-based parsers to substring, this
isn't quite as performant as working directly with the
substring-UTF8View boundary, nor is it as discoverable.

This commit introduces `FromUTF8View`, which is equivalent to many of
the ad hoc `Substring*Parser` wrappers previously defined in this
library, like `SubstringBoolParser`, `SubstringDoubleParser`,
`SubstringIntParser`, etc.

As such, this commit also deprecates those ad hoc parsers for operators
that simply wrap the `UTF8View`-based parser in `FromUTF8View`.

Finally, this commit introduces some more initializer overloads for
`Substring` and `UTF8View` specifically, which are some of the most
common inputs many of the parsers in this library works with. These
initializers should help cut down on some angle bracket intensity.