Prefix should eagerly consume
We had a small bug in how the `Prefix` parser was consuming upon
failure: it would always consume the minimum number of elements, which
is more than the number of elements it actually consumed. This would
produce bad error messages that point to the minimum successful position
and _not_ the failure position.