Downstream ParseBuilder should use Downstream.Input as Input (#291)
* Downstream ParseBuilder should use Downstream.Input as Input

The ParseBuilder is for Downstream so it should use it’s input. Or Upstream.Output as pipe runs the downstream parser on the output of the upstream parser. Which is why Upstream.Output == Downstream.Input.

With the 5.8 changes pipe uses the wrong input here. This only causes issues, if Upstream and Downstream have a different Input type.

* Pipe - Use Upstream.Output instead of Downstream.Input